refactor: 체스 보드와 체스 보드 초기화 로직 분리
