+++ src/main/java/chess/domain/board/ChessBoard.java
@@ -1,34 +1,26 @@
 package chess.domain.board;
 
-import chess.domain.piece.Bishop;
 import chess.domain.piece.Blank;
 import chess.domain.piece.King;
-import chess.domain.piece.Knight;
-import chess.domain.piece.Pawn;
 import chess.domain.piece.Piece;
-import chess.domain.piece.Queen;
-import chess.domain.piece.Rook;
 import chess.domain.result.Pieces;
 import chess.domain.result.ResultDto;
 import chess.domain.result.Score;
 import chess.domain.state.TeamColor;
-import java.util.ArrayList;
 import java.util.HashMap;
-import java.util.LinkedHashMap;
-import java.util.List;
 import java.util.Map;
 
 public class ChessBoard {
 
-    private static final int MAX_NUM_PIECE = 16;
-    private static final int ROW_WHITE_START = 1;
-    private static final int ROW_WHITE_END = 2;
+    public static final int MAX_NUM_PIECE = 16;
+    public static final int ROW_WHITE_START = 1;
+    public static final int ROW_WHITE_END = 2;
     public static final char COLUMN_FIRST = 'a';
     public static final char COLUMN_LAST = 'h';
     public static final int ROW_FIRST = 1;
     public static final int ROW_LAST = 8;
-    private static final int ROW_BLACK_START = 7;
-    private static final int ROW_BLACK_END = 8;
+    public static final int ROW_BLACK_START = 7;
+    public static final int ROW_BLACK_END = 8;
     public static final int ROW_WHITE_PAWN_LINE = 2;
     public static final int ROW_BLACK_PAWN_LINE = 7;
     private static final String EXCEPTION_MOVE = "잘못된 이동입니다.";
@@ -41,23 +33,11 @@ public class ChessBoard {
     private final Pieces blackPieces;
     private boolean gameStatus = true;
 
-    //TODO : init 분리
     public ChessBoard() {
-        this.chessBoard = new LinkedHashMap<>();
-        initBoard();
-        this.whitePieces = initWhitePieces();
-        this.blackPieces = initBlackPieces();
-    }
-
-    private Pieces initWhitePieces() {
-        List<Piece> whitePieces = new ArrayList<>(MAX_NUM_PIECE);
-        for (int row = ROW_WHITE_START; row <= ROW_WHITE_END; row++) {
-            for (char column = COLUMN_FIRST; column <= COLUMN_LAST; column++) {
-                String piecePositionName = createPiecePositionName(row, column);
-                whitePieces.add(piece(Position.valueOf(piecePositionName)));
-            }
-        }
-        return new Pieces(whitePieces);
+        InitChessBoard initChessBoard = new InitChessBoard();
+        this.chessBoard = initChessBoard.getChessBoard();
+        this.whitePieces = initChessBoard.getWhitePieces();
+        this.blackPieces = initChessBoard.getBlackPieces();
     }
 
     public static String createPiecePositionName(int row, char column) {
@@ -66,77 +46,6 @@ public class ChessBoard {
 
     private Piece piece(Position position) {
         return chessBoard.get(position);
-    }
-
-    private Pieces initBlackPieces() {
-        List<Piece> blackPieces = new ArrayList<>(16);
-        for (int row = ROW_BLACK_START; row <= ROW_BLACK_END; row++) {
-            for (char column = COLUMN_FIRST; column <= COLUMN_LAST; column++) {
-                String piecePositionName = createPiecePositionName(row, column);
-                blackPieces.add(piece(Position.valueOf(piecePositionName)));
-            }
-        }
-        return new Pieces(blackPieces);
-    }
-
-    private void initBoard() {
-        for (int row = ROW_FIRST; row <= ROW_LAST; row++) {
-            for (char column = COLUMN_FIRST; column <= COLUMN_LAST; column++) {
-                String boardPosition = createPiecePositionName(row, column);
-                Piece piece = Blank.INSTANCE;
-                pieceOnChessBoard(boardPosition, piece);
-            }
-        }
-        initPawnLine();
-        initUniquePieceLine();
-    }
-
-    private Piece pieceOnChessBoard(String boardPosition, Piece piece) {
-        return chessBoard.put(Position.valueOf(boardPosition), piece);
-    }
-
-    private Piece pieceOnChessBoard(Position boardPosition, Piece piece) {
-        return chessBoard.put(boardPosition, piece);
-    }
-
-    private void initPawnLine() {
-        for (char column = COLUMN_FIRST; column <= COLUMN_LAST; column++) {
-            String boardPosition = createPiecePositionName(ROW_WHITE_PAWN_LINE, column);
-            Piece pawn = new Pawn(TeamColor.WHITE, Position.valueOf(boardPosition));
-            inputPiece(boardPosition, pawn);
-        }
-
-        for (char column = COLUMN_FIRST; column <= COLUMN_LAST; column++) {
-            String boardPosition = createPiecePositionName(ROW_BLACK_PAWN_LINE, column);
-            Piece pawn = new Pawn(TeamColor.BLACK, Position.valueOf(boardPosition));
-            inputPiece(boardPosition, pawn);
-        }
-    }
-
-    private Piece inputPiece(String boardPosition, Piece piece) {
-        return chessBoard.put(Position.valueOf(boardPosition), piece);
-    }
-
-    private void initUniquePieceLine() {
-        TeamColor color = TeamColor.WHITE;
-        chessBoard.put(Position.valueOf("a1"), new Rook(color, Position.valueOf("a1")));
-        chessBoard.put(Position.valueOf("b1"), new Knight(color, Position.valueOf("b1")));
-        chessBoard.put(Position.valueOf("c1"), new Bishop(color, Position.valueOf("c1")));
-        chessBoard.put(Position.valueOf("d1"), new Queen(color, Position.valueOf("d1")));
-        chessBoard.put(Position.valueOf("e1"), new King(color, Position.valueOf("e1")));
-        chessBoard.put(Position.valueOf("f1"), new Bishop(color, Position.valueOf("f1")));
-        chessBoard.put(Position.valueOf("g1"), new Knight(color, Position.valueOf("g1")));
-        chessBoard.put(Position.valueOf("h1"), new Rook(color, Position.valueOf("h1")));
-
-        color = TeamColor.BLACK;
-        chessBoard.put(Position.valueOf("a8"), new Rook(color, Position.valueOf("a8")));
-        chessBoard.put(Position.valueOf("b8"), new Knight(color, Position.valueOf("b8")));
-        chessBoard.put(Position.valueOf("c8"), new Bishop(color, Position.valueOf("c8")));
-        chessBoard.put(Position.valueOf("d8"), new Queen(color, Position.valueOf("d8")));
-        chessBoard.put(Position.valueOf("e8"), new King(color, Position.valueOf("e8")));
-        chessBoard.put(Position.valueOf("f8"), new Bishop(color, Position.valueOf("f8")));
-        chessBoard.put(Position.valueOf("g8"), new Knight(color, Position.valueOf("g8")));
-        chessBoard.put(Position.valueOf("h8"), new Rook(color, Position.valueOf("h8")));
     }
 
     public Map<Position, Piece> getChessBoard() {
@@ -163,6 +72,10 @@ public class ChessBoard {
         if (endPiece instanceof King) {
             gameStatus = false;
         }
+    }
+
+    private Piece pieceOnChessBoard(Position boardPosition, Piece piece) {
+        return chessBoard.put(boardPosition, piece);
     }
 
     private void killPiece(Position start, Position end) {
